requirements.txt
requests
python-dotenv
logging
pandas
pyarrow
matplotlib
seaborn
